move codes to tts module
